Split style.h into finer-grained headers: spec.h, value.h, stylesheet.h (#481)
