+++ libs/vgc/style/CMakeLists.txt
@@ -9,20 +9,24 @@ vgc_add_library(style
         logcategories.h
         metrics.h
         parse.h
+        spec.h
         strings.h
         stylableobject.h
-        style.h
+        stylesheet.h
         token.h
         types.h
+        value.h
 
     CPP_FILES
         logcategories.cpp
         parse.cpp
+        spec.cpp
         strings.cpp
         stylableobject.cpp
-        style.cpp
+        stylesheet.cpp
         token.cpp
         types.cpp
+        value.cpp
 
     COMPILE_DEFINITIONS
 )
